+++ modules/net-vpc-firewall-yaml/outputs.tf
@@ -18,7 +18,7 @@ output "ingress_allow_rules" {
   description = "Ingress rules with allow blocks."
   value = [
     for rule in google_compute_firewall.rules :
-    rule.name if rule.direction == "INGRESS" && length(rule.allow) > 0
+    rule if rule.direction == "INGRESS" && length(rule.allow) > 0
   ]
 }
 
@@ -26,7 +26,7 @@ output "ingress_deny_rules" {
   description = "Ingress rules with deny blocks."
   value = [
     for rule in google_compute_firewall.rules :
-    rule.name if rule.direction == "INGRESS" && length(rule.deny) > 0
+    rule if rule.direction == "INGRESS" && length(rule.deny) > 0
   ]
 }
 
@@ -34,7 +34,7 @@ output "egress_allow_rules" {
   description = "Egress rules with allow blocks."
   value = [
     for rule in google_compute_firewall.rules :
-    rule.name if rule.direction == "EGRESS" && length(rule.allow) > 0
+    rule if rule.direction == "EGRESS" && length(rule.allow) > 0
   ]
 }
 
@@ -42,6 +42,6 @@ output "egress_deny_rules" {
   description = "Egress rules with allow blocks."
   value = [
     for rule in google_compute_firewall.rules :
-    rule.name if rule.direction == "EGRESS" && length(rule.deny) > 0
+    rule if rule.direction == "EGRESS" && length(rule.deny) > 0
   ]
 }